+++ Runtime/Graph/GraphBase/IClonable.cs
@@ -1,0 +1,7 @@
+namespace Talent.Graph
+{
+    public interface IClonable<T>
+    {
+        T GetCopy();
+    }
+}